+++ src/main/java/vn/hactanco/socialnetwork/model/Like.java
@@ -1,0 +1,43 @@
+package vn.hactanco.socialnetwork.model;
+
+import java.time.Instant;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import jakarta.persistence.UniqueConstraint;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Table(name = "likes", uniqueConstraints = { @UniqueConstraint(columnNames = { "user_id", "post_id" }) })
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class Like {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+
+	// user like
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "user_id")
+	private User user;
+
+	// post được like
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "post_id")
+	private Post post;
+
+	@Column(name = "created_at")
+	private Instant createdAt = Instant.now();
+}