+++ src/main/java/vn/hactanco/socialnetwork/model/User.java
@@ -4,6 +4,7 @@ import java.io.Serializable;
 import java.time.Instant;
 import java.util.List;
 
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -64,4 +65,7 @@ public class User implements Serializable {
 
 	@OneToMany(mappedBy = "user", fetch = FetchType.LAZY)
 	private List<Post> posts;
+
+	@OneToMany(mappedBy = "user", cascade = CascadeType.ALL)
+	private List<Like> likes;
 }